[level 2] Title: n^2 배열 자르기, Time: 1.81 ms, Memory: 36.8 MB -BaekjoonHub

+++ 프로그래머스/lv2/87390. n＾2 배열 자르기/n＾2 배열 자르기.cs
@@ -1,16 +1,16 @@
 using System;
 
 public class Solution {
-    public long[] solution(int n, long left, long right) {
+    public int[] solution(int n, long left, long right) {
         
-        long[] answer = new long[right - left + 1];
+        int[] answer = new int[right - left + 1];
         
         for(long i = 0; i < answer.Length; i++)
         {
             long x = (left + i) % n;
             long y = (left + i) / n;
             
-            answer[i] = x > y ? ++x : ++y;
+            answer[i] = (int)(x > y ? ++x : ++y);
         }
         
         return answer;
